Update sortinghat command
When the command is executed, the bot sends a message asking the user
to pick a house. 4 buttons will appear; one for each house. When a
button is clicked, the user is given that house's role. When another
button is clicked, the old house is removed and a new one is applied.