Button interaction
Extracted the button interaction from interactionCreate event. This
makes maintaining the interaction better and allows better scalability.

Fixed bug where the bot wouldn't delete and give new role to the user.
Instead, it'll send an ephemerel message to ask the user contact
server admin to fix the issue. The fix is to simply reorder role
priority.